Make IDs long

+++ src/main/scala/shortenurl/domain/model/Model.scala
@@ -1,4 +1,4 @@
 package shortenurl.domain.model
 
-case class User(id: Int, token: String)
-case class Folder(id: Int, uid: Int, title: String)
+case class User(id: Long, token: String)
+case class Folder(id: Long, uid: Long, title: String)
