+++ src/main/scala/shortenurl/domain/repository/FolderRepositoryComponent.scala
@@ -7,13 +7,13 @@ trait FolderRepositoryComponent { this: FolderTable =>
   val folderRepository: FolderRepository
 
   trait FolderRepository {
-    def listFolders(userId: Int): List[Folder]
+    def listFolders(userId: Long): List[Folder]
   }
 
   class FolderRepositoryImpl extends FolderRepository {
     import profile.simple._
 
-    override def listFolders(uid: Int): List[Folder] = {
+    override def listFolders(uid: Long): List[Folder] = {
       db withSession { implicit session =>
         folders.filter(_.uid === uid).list
       }
